--- components/projects/upload/Toast/types.tsx
@@ -1,4 +1,0 @@
-export type Toast = {
-  isActive: boolean;
-  message: string;
-};
