+++ components/projects/upload/ToastProvider/context.ts
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const ToastContext = createContext({ showToast(message: string) {} });
