+++ components/projects/upload/ToastProvider/types.tsx
@@ -1,0 +1,8 @@
+export type ToastStatus = {
+  isActive: boolean;
+  message: string;
+};
+
+export interface Toast extends ToastStatus {
+  timeoutID: NodeJS.Timeout | null;
+}
